+++ src/status.js
@@ -1,9 +1,17 @@
-import { itemList } from "./index";
 
-export const completed = () => {
-    for (let i = 0; i < itemList.length; i++) {
-        var isComplete = itemList[i].completed;
-    }
-    return isComplete;
-};
-  
+
+const status = () => {
+    const checks = Object.values(document.getElementsByName('status'));
+    
+    checks.forEach(element => {
+        element.addEventListener('click', () => {
+          if (element.checked) {
+            element.parentElement.style.textDecoration = 'underline line-through';
+          } else {
+            element.parentElement.style.textDecoration = 'none';
+          }
+        });
+    });
+}
+
+export { status };
